Sentiment_API/SentimentAnalysisAPI/requirements.txt
fastapi
uvicorn
sqlalchemy
textblob
pyjwt
pydantic